Stop using the word "potentially"
The word "potentially" implies a certain likelihood - even more so for non-native english speakers, which is not a determination that the script can make just be looking at the version output. That fact of the matter is the updated systems hit this code path, and people continue to get confused (see at least 3 github issues).

We don't want to spread FUD, so lets make this output clearer.